+++ mathics/eval/patterns.py
@@ -1,5 +1,11 @@
+from typing import Optional as OptionalType
+
+from mathics.core.atoms import Integer
 from mathics.core.evaluation import Evaluation
+from mathics.core.expression import Expression
 from mathics.core.pattern import BasePattern, StopGenerator
+from mathics.core.symbols import Symbol
+from mathics.core.systemsymbols import SymbolDefault
 
 
 class _StopGeneratorMatchQ(StopGenerator):
@@ -25,6 +31,36 @@ class Matcher:
         except _StopGeneratorMatchQ:
             return True
         return False
+
+
+def get_default_value(
+    name: str,
+    evaluation: Evaluation,
+    k: OptionalType[int] = None,
+    n: OptionalType[int] = None,
+):
+    """
+    Get the default value associated to a name, and optionally,
+    to a position in the expression.
+    """
+    pos = []
+    if k is not None:
+        pos.append(k)
+    if n is not None:
+        pos.append(n)
+    for pos_len in reversed(range(len(pos) + 1)):
+        # Try patterns from specific to general
+        defaultexpr = Expression(
+            SymbolDefault, Symbol(name), *[Integer(index) for index in pos[:pos_len]]
+        )
+        result = evaluation.definitions.get_value(
+            name, "System`DefaultValues", defaultexpr, evaluation
+        )
+        if result is not None:
+            if result.sameQ(defaultexpr):
+                result = result.evaluate(evaluation)
+            return result
+    return None
 
 
 def match(expr, form, evaluation: Evaluation):
